Update color model: - Changed CorelColorPickerModel to CorelTriangleColorPickerModel - Add new CorelSquareColorPickerModel
